fix: 실패한 날씨 패치 업데이트 중 다시 실패한 경우 모두 삭제해버리는 버그 수정

+++ src/main/java/com/boram/look/service/weather/forecast/ForecastScheduler.java
@@ -56,6 +56,7 @@ public class ForecastScheduler {
             }
 
             List<ForecastDto> dailyList = forecastService.saveShortTermsForecastByRegion(forecastDtos, regionId);
+            if (dailyList.isEmpty()) return;
             forecastCacheService.put(regionId.toString(), dailyList);
             forecastFailureService.removeFailure(regionId); // 성공 시 삭제
         }
